👌 New user-facing addition: Move site preview metadata to share page
