+++ packages/settings/src/constants.ts
@@ -23,7 +23,7 @@ export const defaultSettings = {
     description:
       "Build beautiful conversational forms and embed them directly in your applications without a line of code. Triple your response rate and collect answers that has more value compared to a traditional form.",
     favIconUrl: (viewerBaseUrl: string) => viewerBaseUrl + "/favicon.svg",
-    imageUrl: (viewerBaseUrl: string) => viewerBaseUrl + "/site-preview.png",
+    imageUrl: (viewerBaseUrl: string) => viewerBaseUrl + "/site-preview.jpeg",
   },
 } as const;
