+++ packages/ui/src/components/Textarea.tsx
@@ -15,7 +15,7 @@ export const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
       {...props}
       data-slot="textarea"
       className={cn(
-        "rounded-md border bg-gray-1 max-w-sm w-full px-4 py-2.5 outline-hidden ring-orange-7 transition-[color,border-color,box-shadow] focus:border-orange-8 focus:ring-2 field-sizing-content",
+        "rounded-md border bg-gray-1 w-full px-4 py-2.5 outline-hidden ring-orange-7 transition-[color,border-color,box-shadow] focus:border-orange-8 focus:ring-2 field-sizing-content",
         size === "none" ? "min-h-0" : "min-h-20.5 resize-y",
         className,
       )}
